Don't use global trl methods, load them using commonjs
also move commonjs assets above others, so we can define stuff in Kwf. to be used later (needed for Trl)

and also remove default dependency for trl, is now loaded with Kwf

and finally, leave trl/trlc/trlKwf methods as they are in the source code to be able to alias them in commonjs
