+++ Kwc/Menu/Mobile/Component.defer.js
@@ -1,6 +1,7 @@
 var onReady = require('kwf/on-ready');
 var _ = require('underscore');
 var $ = require('jquery');
+var trlKwf = require('kwf/trl').trlKwf;
 
 onReady.onRender('.cssClass', function mobileMenu(el, config) {
     var slideDuration = 400;
