+++ Kwf_js/Trl.js
@@ -1,41 +1,9 @@
-function trl (text, values){
-  if (values == null){
-           return text;
-    } else {
-      if (typeof(values) == 'string' || typeof(values) == 'number') {
-        var temp = values;
-        values = new Array();
-        values.push(temp);
-      }
-      var cnt = 0;
-            values.each(function(value) {
-        text = text.replace(new RegExp('\\{('+cnt+')\\}', 'g'), value);
-        cnt++;
-      });
-           return text;
-       }
-}
-
-function trlp (single, plural, values){
-  if (values == null){
-           return '';
-    } else {
-      if (typeof(values) == 'string' || typeof(values) == 'number') {
-        var temp = values;
-        values = new Array();
-        values.push(temp);
-      }
-
-      if (values[0] == 1){
-        text = single;
-      } else {
-        text = plural;
-      }
-      var cnt = 0;
-            values.each(function(value) {
-        text = text.replace(new RegExp('\\{('+cnt+')\\}', 'g'), value);
-        cnt++;
-      });
-           return text;
-       }
-}
+var t = require('kwf/trl');
+window.trl = t.trl;
+window.trlc = t.trlc;
+window.trlp = t.trlp;
+window.trlcp = t.trlcp;
+window.trlKwf = t.trlKwf;
+window.trlcKwf = t.trlcKwf;
+window.trlpKwf = t.trlpKwf;
+window.trlcpKwf = t.trlcpKwf;
