+++ commonjs/frontend-form/field/file.js
@@ -3,6 +3,7 @@
 var fieldRegistry = require('kwf/frontend-form/field-registry');
 var Field = require('kwf/frontend-form/field/field');
 var kwfExtend = require('kwf/extend');
+var trlKwf = require('kwf/trl').trlKwf;
 
 var File = kwfExtend(Field, {
     initField: function() {
