+++ dependencies.ini
@@ -5,6 +5,7 @@ provider = Kwf_Assets_Provider_Ini
 [dependencies]
 Kwf.files[] = kwf/Kwf_js/Kwf.js
 Kwf.dep[] = KwfExtend
+Kwf.dep[] = KwfTrl
 
 KwfExtend.files[] = kwf/Kwf_js/KwfExtend.js
 
@@ -537,6 +538,9 @@ Frontend.dep[] = FrontendCore
 
 KwfLegacyOnReady.files[] = kwf/Kwf_js/Legacy/OnReady.js
 KwfLegacyOnReady.commonJsEntry = true
+
+KwfTrl.files[] = kwf/Kwf_js/Trl.js
+KwfTrl.commonJsEntry = true
 
 Admin.dep[] = Frontend
 Admin.dep[] = ComponentsAdmin
